Fix download over SSL using IE < 8 and binary file response
