+++ src/Symfony/Component/HttpFoundation/Tests/BinaryFileResponseTest.php
@@ -15,7 +15,7 @@ use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
 
-class BinaryFileResponseTest extends \PHPUnit_Framework_TestCase
+class BinaryFileResponseTest extends ResponseTestCase
 {
     public function testConstruction()
     {
@@ -144,5 +144,10 @@ class BinaryFileResponseTest extends \PHPUnit_Framework_TestCase
             array('/var/www/var/www/files/foo.txt', '/files/=/var/www/', '/files/var/www/files/foo.txt'),
             array('/home/foo/bar.txt', '/files/=/var/www/,/baz/=/home/foo/', '/baz/bar.txt'),
         );
+    }
+
+    protected function provideResponse()
+    {
+        return new BinaryFileResponse('README.md');
     }
 }
